feat(generic_type_utility): rename package
